Add required Name property to QuickMeal entity
Updated the QuickMeal entity to include a new required property `Name`.
Created a migration class `Schema_QuickMeal_AddName` to add the `Name` column
to the QuickMeal table. Updated the ApplicationDbContextModelSnapshot
to reflect the new schema. Modified the QuickMeal constructor to require
the `name` parameter, and added `required` modifiers for both `Name`
and `UserId` properties.